New state management system.
This is a big overhaul, but it’s much better. It uses a redux-inspired
state management system that makes writing systems and manipulating
state *very* easy.

New systems have the type signature of:

    system :: State -> Action -> State

(lol haskell)

The inspiration of this change came from trying to implement the action
system — it ended up becoming really complicated. The games
data-structure was evolving outside of the CES and things were becoming
more and more hockey (multiple queues). These changes have *vastly*
simplified the system.