Remove flair from the game.
This… might be controversial but I feel that while was a fun feature
for Railscamp, it puts the system into a slippery-slope of extending
the command namespace for any such reason. I’d like to figure out how
to do customizations of bots but want to think about how to do that
holistically.